Split initialize package
